feat (modele) : ajout d'une vérification lors de la suppression d'un modele : seul un admin peut le faire

+++ loc-mns/src/main/java/com/mns/cda/locmns/controller/ModeleController.java
@@ -5,8 +5,11 @@ import com.fasterxml.jackson.annotation.JsonView;
 import com.mns.cda.locmns.dto.CreateModeleDto;
 import com.mns.cda.locmns.dto.UpdateModeleDto;
 import com.mns.cda.locmns.model.Modele;
+import com.mns.cda.locmns.model.Role;
+import com.mns.cda.locmns.model.RoleNom;
 import com.mns.cda.locmns.security.IsAdmin;
 import com.mns.cda.locmns.security.IsUser;
+import com.mns.cda.locmns.security.UtilisateurDetails;
 import com.mns.cda.locmns.service.ModeleService;
 import com.mns.cda.locmns.view.ModeleView;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -15,6 +18,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -62,7 +66,13 @@ public class ModeleController {
 
     @DeleteMapping("delete/{id}")
     @IsAdmin
-    public ResponseEntity<Void> delete(@PathVariable int id) {
+    public ResponseEntity<Void> delete(@AuthenticationPrincipal UtilisateurDetails utilisateurDetails,
+            @PathVariable int id) {
+
+        if(utilisateurDetails.getUtilisateur().getRoles().stream().noneMatch(role -> role.getRole() == RoleNom.ADMIN)){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
